refactor: split survey views from feedback
